Print string function doc
