+++ _putchar.c
@@ -10,25 +10,3 @@ int _putchar(char c)/*Function helper*/
 {
 	return (write(1, &c, 1));
 }
-
-/**
- *print_string - Print a string of characters
- *@str: the string to print
- *
- *Return: number of characters printed
- */
-int print_string(char *str)
-{
-	int count = 0;
-
-	if (str == NULL)
-		str = "(null)"; /* Handle NULL string */
-
-	while (*str)
-	{
-		count += _putchar(*str);
-		str++;
-	}
-	return (count);
-}
-
